+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * 'O' shortcut to open options overlay window not working.
   * Options overlay window not working for line segments.
   * Options should be added to copy of original element.
+  * ID option update failing.
 * Infrastructure maintenance:
   * flutter pub upgrade --major-versions.
   * Created onDeselectAll() and onSelectAll() state and actuator methods.